Fix: 응원하기, 팀원 깨우기 API 컨트롤러 변경 (#64) - FieldController -> UserFieldController
