+++ src/main/java/com/dnd/Exercise/domain/field/controller/FieldController.java
@@ -287,21 +287,4 @@ public class FieldController {
         FindFieldResultRes result = fieldService.findFieldResult(user, fieldId);
         return ResponseDto.ok(result);
     }
-
-    @ApiOperation(value = "팀원 깨우기 💡", notes = "2시간에 한 번만 가능하도록")
-    @PostMapping("/alert/{id}")
-    public ResponseEntity<String> alertMembers(
-            @Parameter(description = "필드 ID") @PathVariable("id") Long id){
-        return ResponseDto.ok("팀원 꺠우기 완료");
-    }
-
-
-    @ApiOperation(value = "응원하기 💡", notes = "2시간에 한 번만 가능하도록")
-    @PostMapping("/cheer/{id}")
-    public ResponseEntity<String> cheerMember(
-            @AuthenticationPrincipal User user,
-            @Parameter(description = "유저 ID") @PathVariable("id") Long id){
-        fieldService.cheerMember(user ,id);
-        return ResponseDto.ok("응원하기 완료");
-    }
 }
